job-matcher/backend/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.22.0
python-dotenv==1.0.1
sqlalchemy==2.0.21
alembic==1.11.1
pydantic==1.10.12
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.1.0
langchain==0.1.16
langchain-google-genai==0.0.9
chromadb==0.4.22
pypdf==3.11.0
python-docx==0.8.11
requests==2.31.0
